+++ examplepatches/alan7383/sofatime-patches/patches/src/main/kotlin/app/sofatime/patches/PremiumPatch.kt
@@ -2,15 +2,13 @@ package app.sofatime.patches
 
 import app.morphe.patcher.extensions.InstructionExtensions.addInstructions
 import app.morphe.patcher.patch.bytecodePatch
-import app.sofatime.fingerprints.isPremiumFingerprint
-import app.sofatime.fingerprints.isPremiumPurchasedFingerprint
-import app.sofatime.fingerprints.isPurchasedFingerprint
+
 import app.sofatime.shared.Constants.SOFATIME_COMPATIBILITY
 
 @Suppress("unused")
 val premiumPatch = bytecodePatch(
-    name = "SofaTime Premium",
-    description = "Unlocks all premium features.",
+    name = "Unlock premium features",
+    description = "Unlocks all premium capabilities.",
     default = true,
 ) {
     compatibleWith(SOFATIME_COMPATIBILITY)
@@ -25,14 +23,6 @@ val premiumPatch = bytecodePatch(
         )
 
         isPremiumPurchasedFingerprint.method.addInstructions(
-            0,
-            """
-                const/4 v0, 0x1
-                return v0
-            """
-        )
-
-        isPurchasedFingerprint.method.addInstructions(
             0,
             """
                 const/4 v0, 0x1