+++ examplepatches/alan7383/sofatime-patches/patches/src/main/kotlin/app/sofatime/shared/Constants.kt
@@ -11,7 +11,7 @@ object Constants {
         apkFileType = ApkFileType.APKM,
         appIconColor = 0x1E1E2E,
         targets = listOf(
-            AppTarget(version = "1.1.2")
+            AppTarget(version = "1.1.4")
         )
     )
 }