+++ examplepatches/jasonwu1994/Gboard-patches/CHANGELOG.md
@@ -1,3 +1,23 @@
+## [2.2.0](https://github.com/jasonwu1994/Gboard-patches/compare/v2.1.0...v2.2.0) (2026-07-29)
+
+### ✨ New Features
+
+* **Gboard:** add the `Use Bluetooth Microphone` patch.
+
+### ✨ 新功能
+
+* **Gboard:** 新增 `Use Bluetooth Microphone` Patch。
+
+## [2.2.0-dev.1](https://github.com/jasonwu1994/Gboard-patches/compare/v2.1.0...v2.2.0-dev.1) (2026-07-28)
+
+### ✨ New Features
+
+* **Gboard:** add the `Use Bluetooth Microphone` patch.
+
+### ✨ 新功能
+
+* **Gboard:** 新增 `Use Bluetooth Microphone` Patch。
+
 ## [2.1.0](https://github.com/jasonwu1994/Gboard-patches/compare/v2.0.0...v2.1.0) (2026-07-26)
 
 ### ✨ New Features